tests/wpt/tests/html/canvas/offscreen/line-styles/2d.line.miter.valid.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.line.miter.valid</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.line.miter.valid</h1>
<p class="desc">Setting miterLimit to valid values works</p>


<script>
var t = async_test("Setting miterLimit to valid values works");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  ctx.miterLimit = 1.5;
  _assertSame(ctx.miterLimit, 1.5, "ctx.miterLimit", "1.5");

  ctx.miterLimit = "1e1";
  _assertSame(ctx.miterLimit, 10, "ctx.miterLimit", "10");

  ctx.miterLimit = 1/1024;
  _assertSame(ctx.miterLimit, 1/1024, "ctx.miterLimit", "1/1024");

  ctx.miterLimit = 1000;
  _assertSame(ctx.miterLimit, 1000, "ctx.miterLimit", "1000");
  t.done();

});
</script>
